Use less accurate floating point operations instead

+++ src/distribution.rs
@@ -5,7 +5,7 @@ use rand::{
 };
 
 #[derive(Clone, Debug)]
-pub struct Categorical<T>(Vec<f64>, WeightedIndex<f64>, Vec<T>);
+pub struct Categorical<T>(Vec<f32>, WeightedIndex<f32>, Vec<T>);
 
 impl<T: PartialEq> PartialEq for Categorical<T> {
     fn eq(&self, r: &Self) -> bool {
@@ -20,7 +20,7 @@ impl<T> Categorical<T> {
     }
 
     #[inline]
-    pub fn probs<'a>(&'a self) -> &'a Vec<f64> {
+    pub fn probs<'a>(&'a self) -> &'a Vec<f32> {
         &self.0
     }
 
@@ -40,7 +40,7 @@ impl<T> Categorical<T> {
     }
 
     #[inline]
-    pub fn sample_ref_pair_rng<'a, R: Rng>(&'a self, rng: &mut R) -> (f64, &'a T) {
+    pub fn sample_ref_pair_rng<'a, R: Rng>(&'a self, rng: &mut R) -> (f32, &'a T) {
         let idx = self.1.sample(rng);
         (self.0[idx], &self.2[idx])
     }
@@ -48,31 +48,31 @@ impl<T> Categorical<T> {
     pub fn uniform<IT: Into<Vec<T>>>(items: IT) -> Self {
         let is: Vec<T> = items.into();
         let l = is.len();
-        Self::new(vec![1.0 / (l as f64); l], is)
+        Self::new(vec![1.0 / (l as f32); l], is)
     }
 
-    pub fn new_normalized<IT: Into<Vec<T>>, IP: Into<Vec<f64>>>(probs: IP, items: IT) -> Self {
-        let mut ps: Vec<f64> = probs.into();
-        let s: f64 = ps.iter().sum();
+    pub fn new_normalized<IT: Into<Vec<T>>, IP: Into<Vec<f32>>>(probs: IP, items: IT) -> Self {
+        let mut ps: Vec<f32> = probs.into();
+        let s: f32 = ps.iter().sum();
         ps.iter_mut().for_each(|p| {
             *p = *p / s;
         });
         Self::new(ps, items)
     }
 
-    pub fn new<IT: Into<Vec<T>>, IP: Into<Vec<f64>>>(probs: IP, items: IT) -> Self {
-        let ps: Vec<f64> = probs.into();
+    pub fn new<IT: Into<Vec<T>>, IP: Into<Vec<f32>>>(probs: IP, items: IT) -> Self {
+        let ps: Vec<f32> = probs.into();
         let is: Vec<T> = items.into();
         assert_eq!(ps.len(), is.len());
-        debug_assert!((ps.iter().sum::<f64>() - 1.0) < 1e-3);
+        debug_assert!((ps.iter().sum::<f32>() - 1.0) < 1e-3);
         let wi = WeightedIndex::new(&ps).expect("invalid distribution");
         Categorical(ps, wi, is)
     }
 
     /*
-    pub fn epsilon_smooth(&mut self, epsilon: f64) {
+    pub fn epsilon_smooth(&mut self, epsilon: f32) {
         assert!(epsilon >= 0.0 && epsilon <= 1.0);
-        self.0.iter_mut().for_each(|mut pr| *pr = *pr * (1.0 - epsilon) + epsilon / self.2.len() as f64);
+        self.0.iter_mut().for_each(|mut pr| *pr = *pr * (1.0 - epsilon) + epsilon / self.2.len() as f32);
         self.1 = WeightedIndex::new(self.0.clone()).expect("invalid distribution");
     }*/
 }
@@ -89,15 +89,15 @@ impl<T: Clone> Categorical<T> {
     }
 
     #[inline]
-    pub fn sample_and_prob<'a, R: Rng>(&'a self, rng: &mut R) -> (T, f64) {
+    pub fn sample_and_prob<'a, R: Rng>(&'a self, rng: &mut R) -> (T, f32) {
         let idx = self.1.sample(rng);
         (self.2[idx].clone(), self.0[idx])
     }
 }
 
-pub fn sample_weighted<R: Rng>(ps: &[f64], rng: &mut R) -> usize {
-    debug_assert!((ps.iter().sum::<f64>() - 1.0).abs() < 1e-6);
-    let mut s: f64 = rng.sample(rand::distributions::Standard);
+pub fn sample_weighted<R: Rng>(ps: &[f32], rng: &mut R) -> usize {
+    debug_assert!((ps.iter().sum::<f32>() - 1.0).abs() < 1e-6);
+    let mut s: f32 = rng.sample(rand::distributions::Standard);
     for (i, p) in ps.iter().enumerate() {
         s -= p;
         if s < 0.0 {
